+++ napari/layers/labels/labels.py
@@ -8,7 +8,7 @@ from scipy import ndimage as ndi
 from ..image import Image
 from ...util.colormaps import colormaps
 from ...util.event import Event
-from ...util.misc import interpolate_coordinates
+from .labels_utils import interpolate_coordinates
 from ...util.status_messages import format_float
 from ._constants import Mode
