+++ napari/layers/labels/labels_utils.py
@@ -1,0 +1,34 @@
+import numpy as np
+
+
+def interpolate_coordinates(old_coord, new_coord, brush_size):
+    """Interpolates coordinates depending on brush size.
+
+    Useful for ensuring painting is continuous in labels layer.
+
+    Parameters
+    ----------
+    old_coord : np.ndarray, 1x2
+        Last position of cursor.
+    new_coord : np.ndarray, 1x2
+        Current position of cursor.
+    brush_size : float
+        Size of brush, which determines spacing of interpolation.
+
+    Returns
+    ----------
+    coords : np.array, Nx2
+        List of coordinates to ensure painting is continuous
+    """
+    num_step = round(
+        max(abs(np.array(new_coord) - np.array(old_coord))) / brush_size * 4
+    )
+    coords = [
+        np.linspace(old_coord[i], new_coord[i], num=num_step + 1)
+        for i in range(len(new_coord))
+    ]
+    coords = np.stack(coords).T
+    if len(coords) > 1:
+        coords = coords[1:]
+
+    return coords
